// web/src/api/generated/types.gen.ts
export type ContentResponse = {
    /**
     * Id
     */
    id: string;
    /**
     * Source Link
     */
    source_link: string | null;
    /**
     * Source Author
     */
    source_author: string | null;
    /**
     * Source Origin
     */
    source_origin: string | null;
    /**
     * Source Format
     */
    source_format: string | null;
    /**
     * Summary
     */
    summary: {
        [key: string]: unknown;
    };
    /**
     * Created At
     */
    created_at: string;
};
// ...
export type CourseResponse = {
    /**
     * Id
     */
    id: string;
    /**
     * Title
     */
    title: string;
    /**
     * Description
     */
    description: string;
    /**
     * Objective
     */
    objective: string;
    /**
     * Genre
     */
    genre: string;
    /**
     * Difficulty Level
     */
    difficulty_level: string;
    /**
     * Estimated Duration Seconds
     */
    estimated_duration_seconds: number;
    /**
     * Created At
     */
    created_at: string;
};